backend/requirements.txt
flask
flask_sqlalchemy
pymysql
python-dotenv
flask-cors
cryptography
flasgger